[REF] Move DGII WS to dgii_tools, and remove uneeded JS
