+++ ncf_manager/test/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-from .import test_ncf_manager
+from . import test_ncf_manager
